fix: resolve CI failures (fmt, clippy, missing test fields)
- Run cargo fmt on all files (legs.rs, leg_persistence.rs, lib.rs,
  auth_handlers.rs, oidc.rs, telemetry.rs)
- Add StubLegRepository to activity handler tests for the new leg_repo
  field in ActivityAppState
- Fix unnecessary cast clippy warning in leg_commands.rs test
- All checks pass: cargo fmt --check, cargo clippy -W clippy::all,
  cargo test --all

+++ crates/app/src/recorded_activity/leg_commands.rs
@@ -362,7 +362,7 @@ mod tests {
 
         // Add max legs
         for i in 1..=MAX_LEGS_PER_ACTIVITY {
-            let date = NaiveDate::from_ymd_opt(2024, 7, i as u32).unwrap();
+            let date = NaiveDate::from_ymd_opt(2024, 7, i).unwrap();
             add_leg(activity_id, None, date, None, None, &repo)
                 .await
                 .unwrap();
